requirements.txt
numpy
tqdm
pillow
torch
transformers
pytest